Add a native event delegate to the Window class
This delegate is executed when the native window events are received and before we do any processing of them. Delegate has bool return type and it can be used by user to signal if we should ignore this event after the delegate or if we should process it. Return true if you don't want us to process it further.