Refactor cursor mode in Window class
Cursor now has three modes: Normal, Hidden and Infinite.

Normal relies on WM_MOUSEMOVE events, cursor is not locked to the window and is visible. Same as cursor in standard windows applications.

Hidden very similar to the Normal mode but the cursor is not visible.

Infinite the cursor is invisible and locked to the app window. After every change it is centered and window only generates relative mouse movement events based on raw mouse input data. Used for FPS camera.

Input system now has a function for submiting relative mouse movement which simplifies the logic for event dispatch.

Modify FlyCamera to use relative axis change to rotate the camera.