Update Setup.ps1
Switch to using Root field in the drive object to get root of the drive.

Make some of the logs be debug, so that they are not seen by default.

Fix exit condition in the for loop that goes over the drives so that it exits the loop if it finds the file.

Color some of the standard output.